Some rewriting and code simplifcation
